update examples (#83)
As of last week, v2 release candidate was dropped and is no longer experimental. Ref: https://github.com/solana-labs/solana-web3.js/releases
Looks like with the release, a function was renamed. We are replacing the `createRpcApi` import with `createJsonRpcApi` and associated uses. Also one of the params changed as well.